+++ azurerm/internal/services/web/app_service.go
@@ -12,6 +12,7 @@ import (
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/tf/suppress"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/tf/validation"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/utils"
+	msivalidate "github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/services/msi/validate"
 )
 
 func schemaAppServiceAadAuthSettings() *pluginsdk.Schema {
@@ -490,6 +491,19 @@ func schemaAppServiceSiteConfig() *pluginsdk.Schema {
 					Type:     pluginsdk.TypeString,
 					Optional: true,
 				},
+
+				"acr_use_managed_identity_creds": {
+					Type:     pluginsdk.TypeBool,
+					Optional: true,
+					Default:  false,
+				},
+
+				"acr_user_managed_identity_id": {
+					Type:         pluginsdk.TypeString,
+					ValidateFunc: msivalidate.UserAssignedIdentityID,
+					Optional:     true,
+				},
+
 			},
 		},
 	}
@@ -818,6 +832,16 @@ func schemaAppServiceDataSourceSiteConfig() *pluginsdk.Schema {
 							},
 						},
 					},
+				},
+
+				"acr_use_managed_identity_creds": {
+					Type:     pluginsdk.TypeBool,
+					Computed: true,
+				},
+
+				"acr_user_managed_identity_id": {
+					Type:     pluginsdk.TypeString
+					Computed: true,
 				},
 			},
 		},
@@ -1773,6 +1797,14 @@ func expandAppServiceSiteConfig(input interface{}) (*web.SiteConfig, error) {
 		siteConfig.AutoSwapSlotName = utils.String(v.(string))
 	}
 
+	if v, ok := config["acr_use_managed_identity_creds"]; ok {
+		siteConfig.AcrUseManagedIdentityCreds = utils.Bool(v.(bool))
+	}
+
+	if v, ok := config["acr_user_managed_identity_id"]; ok {
+		siteConfig.AcrUserManagedIdentityID = utils.String(v.(string))
+	}
+
 	return siteConfig, nil
 }
 
@@ -1882,6 +1914,14 @@ func flattenAppServiceSiteConfig(input *web.SiteConfig) []interface{} {
 
 	if input.AutoSwapSlotName != nil {
 		result["auto_swap_slot_name"] = *input.AutoSwapSlotName
+	}
+
+	if input.AcrUseManagedIdentityCreds != nil {
+		result["acr_use_managed_identity_creds"] = *input.AcrUseManagedIdentityCreds
+	}
+
+	if input.AcrUserManagedIdentityID != nil {
+		result["acr_user_managed_identity_id"] = *input.AcrUserManagedIdentityID
 	}
 
 	return append(results, result)